# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation for e-commerce",
        "contact": {
            "email": "delvinnj02@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "operationId": "userLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logout user",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/collections": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "Get list of collections",
                "operationId": "getCollections",
                "parameters": [
                    {
                        "name": "page_limit",
                        "in": "query",
                        "description": "The number of collections to return per page, the default count is 50",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "products",
                        "in": "query",
                        "description": "The number of products to return per page, the default count is 50",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Collections"
                ],
                "summary": "Create a new collection",
                "operationId": "createCollection",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "htmlContent"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "htmlContent": {
                                        "description": "Product content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/collections/{id}": {
            "put": {
                "tags": [
                    "Collections"
                ],
                "summary": "Update collection",
                "operationId": "updateCollection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collections id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "htmlContent"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "htmlContent": {
                                        "description": "Product content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Collections"
                ],
                "summary": "Delete a collection",
                "operationId": "deleteSingleCollection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "page_limit",
                        "in": "query",
                        "description": "The number of products to return per page, the default count is 50",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "collections",
                        "in": "query",
                        "description": "Including collections",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create new product",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "htmlContent",
                                    "productType",
                                    "vendor"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "htmlContent": {
                                        "description": "Product content",
                                        "type": "string"
                                    },
                                    "productType": {
                                        "description": "Product type",
                                        "type": "string",
                                        "enum": [
                                            "Food",
                                            "Drinks",
                                            "Electronic"
                                        ]
                                    },
                                    "vendor": {
                                        "description": "Vendor name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product details",
                "operationId": "getSingleProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "collections",
                        "in": "query",
                        "description": "Including collections",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "htmlContent",
                                    "productType",
                                    "vendor"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "htmlContent": {
                                        "description": "Product content",
                                        "type": "string"
                                    },
                                    "productType": {
                                        "description": "Product type",
                                        "type": "string",
                                        "enum": [
                                            "Food",
                                            "Drinks",
                                            "Electronic"
                                        ]
                                    },
                                    "vendor": {
                                        "description": "Vendor name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "deleteSingleProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer *********************)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}